backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
Pillow
stegano
uuid
python-magic